get phymem by myself

+++ data_type.h
@@ -12,7 +12,7 @@ typedef struct
 typedef struct
 {
     int pid;// 目标进程
-    unsigned int addr;// 读取的地址
+    unsigned long addr;// 读取的地址
     void * buf_read_addr;// 用户空间读取的缓冲区指针
     int size;// 读取字节数
 }memread_t;
